rag_pipeline/requirements.txt
llama-index>=0.10.0
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-readers-file
llama-index-vector-stores-chroma
pypdf2
python-dotenv
chromadb
sentence-transformers
torch
openai>=1.0.0
numpy
pathlib
typing-extensions
